Continue abstracting scatter plots, apply default to single-gene view
