+++ app/views/site/_expression_plots_view.html.erb
@@ -13,11 +13,11 @@
 </div>
 <div class="tab-pane expression-plot" id="scatter-tab" role="tabpanel">
   <div class="panel panel-default no-top-border">
-    <div id="expression-scatter-plots" class="panel-collapse collapse in">
+    <div id="scatter-plots" class="panel-collapse collapse in">
       <div class="panel-body">
         <div class="row" id="scatter-wrapper">
           <div class="col-md-7">
-            <div id="expression-scatter-plot"></div>
+            <div id="scatter-plot"></div>
           </div>
           <div class="col-md-5">
             <div id="reference-plot"></div>
