+++ app/views/site/_study_visualize.html.erb
@@ -20,7 +20,7 @@
         <% if @study.can_visualize_clusters? %>
           <div class="tab-pane active" id="plots-tab" role="tabpanel">
             <div class="panel panel-default no-top-border">
-              <div id="plots" class="panel-collapse collapse in">
+              <div id="scatter-plots" class="panel-collapse collapse in">
                 <div class="panel-body">
                   <div class="col-md-12">
                     <div class="row">
